prevent multiple subscriptions to the same NodeFeature
