+++ custom_components/plugwise_usb/entity.py
@@ -2,12 +2,10 @@
 
 from __future__ import annotations
 
-from collections.abc import Callable
 from dataclasses import dataclass
 import logging
-from typing import Any
 
-from plugwise_usb.api import PUSHING_FEATURES, NodeFeature, NodeInfo
+from plugwise_usb.api import NodeFeature, NodeInfo
 
 from homeassistant.helpers.device_registry import CONNECTION_ZIGBEE, DeviceInfo
 from homeassistant.helpers.entity import EntityDescription
@@ -40,10 +38,8 @@ class PlugwiseUSBEntity(CoordinatorEntity):
         super().__init__(node_duc, context=entity_description.node_feature)
         self.node_duc = node_duc
         self.entity_description = entity_description
-        self.unsubscribe_push_events: Callable[[], None] | None = None
         self._node_info: NodeInfo = node_duc.node.node_info
         self._attr_unique_id = f"{self._node_info.mac}-{entity_description.key}"
-        self._subscribe_to_feature_fn = node_duc.node.subscribe_to_feature_update
         self._via_device = (DOMAIN, str(node_duc.api_stick.mac_stick))
 
     @property
@@ -70,34 +66,11 @@ class PlugwiseUSBEntity(CoordinatorEntity):
     async def async_added_to_hass(self):
         """Subscribe for push updates."""
         await super().async_added_to_hass()
-        if ( 
-            self.entity_description.node_feature in PUSHING_FEATURES
-            and self.entity_description.node_feature in self._node_info.features
-        ):
-            self.unsubscribe_push_events = self._subscribe_to_feature_fn(
-                self.async_push_event,
-                self.entity_description.node_feature,
-            )
-
-    async def async_push_event(self, feature: NodeFeature, state: Any) -> None:
-        """Update data on pushed event."""
-        if self.node_duc is None:
-            _LOGGER.warning(
-                "Unable to push event=%s, state=%s, mac=%s",
-                feature,
-                state,
-                self._node_info.mac,
-            )
-        else:
-            self.node_duc.async_set_updated_data(
-                {
-                    feature: state,
-                }
-            )
+        await self.node_duc.subscribe_nodefeature(
+                 self.entity_description.node_feature
+              )
 
     async def async_will_remove_from_hass(self):
         """Unsubscribe to updates."""
-        if self.unsubscribe_push_events is not None:
-            self.unsubscribe_push_events()
-            self.unsubscribe_push_events = None
+        await self.node_duc.unsubscribe_all_nodefeatures()
         await super().async_will_remove_from_hass()
